% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheet_combine.R
\name{sheet_combine}
\alias{sheet_combine}
\title{Combine all Partner Weekly Data into one tidy dataset}
\usage{
sheet_combine(folderpath_reports, folderpath_output = NULL)
}
\arguments{
\item{folderpath_reports}{what folder are all the partner reports stored in?}

\item{folderpath_output}{what folder would you like to save the combines output dataset in? If no folder is noted, no file will be saved}
}
\description{
Combine all Partner Weekly Data into one tidy dataset
}
\examples{
 \dontrun{
 df_weekly <- sheet_combine("~/Week_23")}
}
