% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheet_identify.R
\name{sheet_identify}
\alias{sheet_identify}
\title{Identify all files and their sheets within}
\usage{
sheet_identify(folderpath)
}
\arguments{
\item{folderpath}{what folder are the weekly files stored in?}
}
\description{
Identify all files and their sheets within
}
\examples{
\dontrun{
sheet_identify("~/project_x") }
}
