% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheet_import.R
\name{sheet_import}
\alias{sheet_import}
\title{Import & adjust each sheet from file}
\usage{
sheet_import(filepath, sheet_name)
}
\arguments{
\item{filepath}{filepath for one of the weekly reports}

\item{sheet_name}{name of the sheet for one of the sheets within the file, eg "HTS_TST"}
}
\description{
Import & adjust each sheet from file
}
\examples{
\dontrun{
  df_parterdata <- sheet_import("~/X Weekly Programme monitoring.xlsx", "HTS_TST")}
}
