% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpm_addmechid.R
\name{wpm_addmechid}
\alias{wpm_addmechid}
\title{Map mechanism ids onto partner names}
\usage{
wpm_addmechid(df)
}
\arguments{
\item{df}{which df to map onto?}
}
\description{
Map mechanism ids onto partner names
}
\examples{
\dontrun{
  df_weekly <- wpm_addmechid(df_weekly_full)}
}
