% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpm_addpds.R
\name{wpm_addpds}
\alias{wpm_addpds}
\title{Add month and fiscal quarter}
\usage{
wpm_addpds(df, date_var)
}
\arguments{
\item{df}{data frame to add week number to}

\item{date_var}{unquoted variable that contains the date}
}
\description{
Add month and fiscal quarter
}
\examples{
\dontrun{
  df_weekly <- wpm_addpds(df_weekly, date)}
}
