% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpm_addtargets.R
\name{wpm_addtargets}
\alias{wpm_addtargets}
\title{Merge targets onto dataframe}
\usage{
wpm_addtargets(df, folderpath_targets)
}
\arguments{
\item{df}{dataframe to add targets onto}

\item{folderpath_targets}{folder path where targets are stored}
}
\description{
Merge targets onto dataframe
}
