% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpm_addweek.R
\name{wpm_addweek}
\alias{wpm_addweek}
\title{Add FY week number}
\usage{
wpm_addweek(df, date_var, fy_start = "2017-10-02")
}
\arguments{
\item{df}{data frame to add week number to}

\item{date_var}{unquoted variable that contains the date}

\item{fy_start}{first Monday of FY, default = "2017-10-02"}
}
\description{
Add FY week number
}
\examples{
\dontrun{
  df_weekly <- wpm_addweek(df_weekly, date)}
}
