% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpm_combine.R
\name{wpm_combine}
\alias{wpm_combine}
\title{Combine all Partner Weekly Data into one tidy dataset}
\usage{
wpm_combine(folderpath_reports, folderpath_sitecoords = NULL,
  folderpath_targets = NULL, folderpath_output = NULL)
}
\arguments{
\item{folderpath_reports}{what folder are all the partner reports stored in?}

\item{folderpath_sitecoords}{what folder is the site coordinates files stored in? If no folder is noted, no coordinates will be merged on}

\item{folderpath_targets}{what folder is the targets file stored in?}

\item{folderpath_output}{what folder would you like to save the combines output dataset in? If no folder is noted, no file will be saved}
}
\description{
Combine all Partner Weekly Data into one tidy dataset
}
\examples{
 \dontrun{
 df_weekly <- wpm_combine("~/Week_23")}
}
