% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpm_combine_monthly.R
\name{wpm_combine_monthly}
\alias{wpm_combine_monthly}
\title{Combine all Partner MONTLY Data into one tidy dataset}
\usage{
wpm_combine_monthly(folderpath_reports, folderpath_output = NULL)
}
\arguments{
\item{folderpath_reports}{what folder are all the partner reports stored in?}

\item{folderpath_output}{what folder would you like to save the combines output dataset in? If no folder is noted, no file will be saved}
}
\description{
Combine all Partner MONTLY Data into one tidy dataset
}
\examples{
 \dontrun{
 df_weekly <- wpm_combine("~/Week_23")}
}
