% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpm_filerefresh.R
\name{wpm_filerefresh}
\alias{wpm_filerefresh}
\title{Remove old files and replace with new}
\usage{
wpm_filerefresh(filepath_zipped, folderpath_reports)
}
\arguments{
\item{filepath_zipped}{what is the full filepath to the zipped file?}

\item{folderpath_reports}{what folder are all the partner reports stored in?}
}
\description{
Remove old files and replace with new
}
\examples{
\dontrun{
 wpm_fileupdate("~/Downloads/drive-download-20180502T170410Z-001.zip", "~/Week_23/reports")}
}
