% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpm_genlinkage.R
\name{wpm_genlinkage}
\alias{wpm_genlinkage}
\title{Create a linkage variable}
\usage{
wpm_genlinkage(df)
}
\arguments{
\item{df}{dataframe to add proxy linkage to}
}
\description{
Create a linkage variable
}
