% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpm_identify.R
\name{wpm_identify}
\alias{wpm_identify}
\title{Identify all files and their sheets within}
\usage{
wpm_identify(folderpath)
}
\arguments{
\item{folderpath}{what folder are the weekly files stored in?}
}
\description{
Identify all files and their sheets within
}
\examples{
\dontrun{
wpm_identify("~/project_x") }
}
