% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpm_import_monthly.R
\name{wpm_import_monthly}
\alias{wpm_import_monthly}
\title{Import & adjust each MONTHLY sheet from file}
\usage{
wpm_import_monthly(filepath, sheet_name)
}
\arguments{
\item{filepath}{filepath for one of the montly reports}

\item{sheet_name}{name of the sheet for one of the sheets within the file, eg "Oct 2017"}
}
\description{
Import & adjust each MONTHLY sheet from file
}
\examples{
\dontrun{
  df_parterdata <- wpm_import("~/ANOVA - 17020_ HTS_POS linkage to TIER.NET.xlsx", "Oct 2017")}
}
