% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpm_map.R
\name{wpm_map}
\alias{wpm_map}
\title{Add coordinates and uids to}
\usage{
wpm_map(df, folderpath_orgunits)
}
\arguments{
\item{df}{dataframe to add coorindates/uids to}

\item{folderpath_orgunits}{what folder are the org units stored in?}
}
\description{
Add coordinates and uids to
}
\examples{
\dontrun{
 df <- wpm_combine(df,"~/WPM/GIS")}
}
