% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpm_getcoords.R
\name{wpm_getcoords}
\alias{wpm_getcoords}
\title{Call site coordinates and hierarchy from DATIM via API}
\usage{
wpm_getcoords(organisationUnit = NA, level = NA, ou_name = NA,
  folderpath_export = NULL)
}
\arguments{
\item{organisationUnit}{what is the UID for the country you would like site coordinates for?}

\item{level}{what is the site level in DATIM, eg 7?}

\item{ou_name}{what is the name of the country? (used for saving)}

\item{folderpath_export}{what is the folderpath where you would like to save the export the file (csv)?}
}
\description{
Adapted from J.Pickering's DATIM Vaidation package - datimvalidation::getOrganisationUnitMap(). Prior to running, you MUST run datimvalidation::loadSecrets()
}
\examples{
\dontrun{
  wpm_getcoords("cDGPF739ZZr", 7, "ZAF", "GIS")}

}
